+++ src/onegov.core/onegov/core/tests/test_custom_json.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from onegov.core.custom_json import dumps, loads
 
 
@@ -10,7 +11,8 @@ def test_custom_json():
         'datetime': dt,
         'date': dt.date(),
         'time': dt.time(),
-        'generator': (x for x in range(0, 4))
+        'generator': (x for x in range(0, 4)),
+        'decimal': Decimal('3.1415926')
     }
 
     text = dumps(data)
@@ -18,6 +20,7 @@ def test_custom_json():
     assert '__time__@12:00:00' in text
     assert '__date__@2015-06-25' in text
     assert '__datetime__@2015-06-25T12:00:00' in text
+    assert '__decimal__@3.1415926' in text
     assert '[0, 1, 2, 3]' in text
 
     data['generator'] = [x for x in range(0, 4)]